hyperv - .env for argocd

+++ ansible-terraform/hyper-v/run.sh
@@ -5,6 +5,13 @@ if [ $# -eq 0 ]; then
   exit 0
 fi
 
+if [ ! -f .env ]; then
+  echo "ERR: .env file missing"
+  exit 3
+fi
+
+. .env
+
 if [ "$1" == "stop" ]; then
   terraform destroy -auto-approve
   exit 0
